requirements.txt
Flask
spotipy
python-dotenv
python-dateutil
uuid